Move Bitcoin tab.
