+++ WalletWasabi.Fluent/Views/Settings/GeneralSettingsView.xaml
@@ -5,8 +5,8 @@
              xmlns:settings="clr-namespace:WalletWasabi.Fluent.ViewModels.Settings"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="WalletWasabi.Fluent.Views.Settings.GeneralSettingsView"
+             x:DataType="settings:SettingsPageViewModel"
              DataContext="{Binding GeneralTab}">
-             <!-- x:DataType="settings:SettingsPageViewModel" -->
   <StackPanel Spacing="20" MinWidth="360" MaxWidth="650" HorizontalAlignment="Left" Orientation="Vertical">
           <StackPanel.Styles>
             <Style Selector="DockPanel > TextBlock">
