+++ WalletWasabi.Fluent/Views/Settings/NetworkSettingsView.xaml
@@ -5,7 +5,8 @@
              xmlns:settings="clr-namespace:WalletWasabi.Fluent.ViewModels.Settings"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:DataType="settings:SettingsPageViewModel"
-             x:Class="WalletWasabi.Fluent.Views.Settings.NetworkSettingsView">
+             x:Class="WalletWasabi.Fluent.Views.Settings.NetworkSettingsView"
+             DataContext="{Binding NetworkTab}">
   <StackPanel Spacing="20" MinWidth="360" MaxWidth="650" HorizontalAlignment="Left" Orientation="Vertical">
     <StackPanel.Styles>
       <Style Selector="DockPanel > TextBlock">
